Feat: 해당 날짜를 클릭하면 데이터들 렌더링
해당 날짜를 클릭하면 달력 아래에 렌더링 되도록 구현함
