+++ components/common/FocusDay/index.tsx
@@ -1,0 +1,22 @@
+import { StyledFocusDay } from './styles';
+import { FocusDayPropsType } from './type';
+
+const FocusDay = ({ focusDay }: FocusDayPropsType) => {
+  return (
+    <StyledFocusDay>
+      <div>
+        {focusDay.year}.{focusDay.month}.{focusDay.date}
+      </div>
+      <div>
+        {focusDay.data.map((scheduleItem: any, scheduleItemIndex: number) => (
+          <div key={scheduleItemIndex}>
+            <div>제목: {scheduleItem.title}</div>
+            <div>내용: {scheduleItem.content}</div>
+          </div>
+        ))}
+      </div>
+    </StyledFocusDay>
+  );
+};
+
+export default FocusDay;
