+++ components/common/FocusDay/styles.tsx
@@ -1,0 +1,7 @@
+import styled from '@emotion/styled';
+
+export const StyledFocusDay = styled.div`
+  width: 70%;
+  margin: 0 auto;
+  border: 1px solid black;
+`;
