+++ components/common/FocusDay/type.ts
@@ -1,0 +1,5 @@
+import { FocusDayType } from '../../Schedule/type';
+
+export type FocusDayPropsType = {
+  focusDay: FocusDayType;
+};
