+++ mocks/mockCalendarData.ts
@@ -3,21 +3,41 @@ export const mockCalendarData = [
     year: '2022',
     month: '8',
     date: '1',
-    data: [{ item: '스택 다시보기', backgroundColor: '#f07e7e' }],
+    data: [
+      {
+        title: '스택 다시보기',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: '#f07e7e',
+      },
+    ],
   },
   {
     year: '2022',
     month: '8',
     date: '3',
-    data: [{ item: '큐 다시보기', backgroundColor: '#7ef0e3' }],
+    data: [
+      {
+        title: '큐 다시보기',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: '#7ef0e3',
+      },
+    ],
   },
   {
     year: '2022',
     month: '8',
     date: '9',
     data: [
-      { item: '프로그래머스 복습', backgroundColor: '#d9f07e' },
-      { item: '백준 1323 복습', backgroundColor: 'skyblue' },
+      {
+        title: '프로그래머스 복습',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: '#d9f07e',
+      },
+      {
+        title: '백준 1323 복습',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: 'skyblue',
+      },
     ],
   },
   {
@@ -25,23 +45,45 @@ export const mockCalendarData = [
     month: '8',
     date: '14',
     data: [
-      { item: '코딩테스트 리뷰', backgroundColor: '#c6aca5' },
-      { item: '백준 5232 복습', backgroundColor: '#aaa5c6' },
+      {
+        title: '코딩테스트 리뷰',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: '#c6aca5',
+      },
+      {
+        title: '백준 5232 복습',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: '#aaa5c6',
+      },
     ],
   },
   {
     year: '2022',
     month: '8',
     date: '19',
-    data: [{ item: '데이터1', backgroundColor: '#a5c6a8' }],
+    data: [
+      {
+        title: '데이터1',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: '#a5c6a8',
+      },
+    ],
   },
   {
     year: '2022',
     month: '8',
     date: '24',
     data: [
-      { item: '데이터1', backgroundColor: 'skyblue' },
-      { item: '데이터2', backgroundColor: '#6165c2' },
+      {
+        title: '데이터1',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: 'skyblue',
+      },
+      {
+        title: '데이터2',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: '#6165c2',
+      },
     ],
   },
   {
@@ -50,10 +92,15 @@ export const mockCalendarData = [
     date: '29',
     data: [
       {
-        item: '데이터asdadajhgsjdajsgasdasdasdasd',
+        title: '데이터asdadajhgsjdajsgasdasdasdasd',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
         backgroundColor: '#9bc261',
       },
-      { item: '데이터2', backgroundColor: '#c2a761' },
+      {
+        title: '데이터2',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: '#c2a761',
+      },
     ],
   },
   {
@@ -62,10 +109,16 @@ export const mockCalendarData = [
     date: '3',
     data: [
       {
-        item: '데이터ㅁㄴ옴ㄴㅇㅎ머ㅗㄴㅇ허뫃ㄴㅇ머농ㅁㄴㅇㅁㄴㅇㅁㅇㅁㄴㅇㅁㄴㅇㅁㄴㅇ',
+        title:
+          '데이터ㅁㄴ옴ㄴㅇㅎ머ㅗㄴㅇ허뫃ㄴㅇ머농ㅁㄴㅇㅁㄴㅇㅁㅇㅁㄴㅇㅁㄴㅇㅁㄴㅇ',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
         backgroundColor: 'purple',
       },
-      { item: '데이터2', backgroundColor: '#239232' },
+      {
+        title: '데이터2',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: '#239232',
+      },
     ],
   },
   {
@@ -73,11 +126,31 @@ export const mockCalendarData = [
     month: '9',
     date: '6',
     data: [
-      { item: '데이터1', backgroundColor: '#61c0c2' },
-      { item: '데이터2', backgroundColor: '#7bc261' },
-      { item: '백준 1323 복습', backgroundColor: 'skyblue' },
-      { item: '백준 1323 복습', backgroundColor: 'skyblue' },
-      { item: '백준 1323 복습', backgroundColor: 'skyblue' },
+      {
+        title: '데이터1',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: '#61c0c2',
+      },
+      {
+        title: '데이터2',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: '#7bc261',
+      },
+      {
+        title: '백준 1323 복습',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: 'skyblue',
+      },
+      {
+        title: '백준 1323 복습',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: 'skyblue',
+      },
+      {
+        title: '백준 1323 복습',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: 'skyblue',
+      },
     ],
   },
   {
@@ -85,8 +158,16 @@ export const mockCalendarData = [
     month: '9',
     date: '11',
     data: [
-      { item: '데이터1', backgroundColor: 'purple' },
-      { item: '데이터2', backgroundColor: '#239232' },
+      {
+        title: '데이터1',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: 'purple',
+      },
+      {
+        title: '데이터2',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: '#239232',
+      },
     ],
   },
   {
@@ -94,29 +175,57 @@ export const mockCalendarData = [
     month: '9',
     date: '15',
     data: [
-      { item: '데이터1', backgroundColor: '#ff00ff' },
-      { item: '데이터2', backgroundColor: '#239232' },
+      {
+        title: '데이터1',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: '#ff00ff',
+      },
+      {
+        title: '데이터2',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: '#239232',
+      },
     ],
   },
   {
     year: '2022',
     month: '9',
     date: '24',
-    data: [{ item: '데이터1', backgroundColor: '#a5c6a8' }],
+    data: [
+      {
+        title: '데이터1',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: '#a5c6a8',
+      },
+    ],
   },
   {
     year: '2022',
     month: '9',
     date: '25',
-    data: [{ item: '스택 다시보기', backgroundColor: '#f07e7e' }],
+    data: [
+      {
+        title: '스택 다시보기',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: '#f07e7e',
+      },
+    ],
   },
   {
     year: '2022',
     month: '9',
     date: '26',
     data: [
-      { item: '프로그래머스 복습', backgroundColor: '#d9f07e' },
-      { item: '백준 1323 복습', backgroundColor: 'skyblue' },
+      {
+        title: '프로그래머스 복습',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: '#d9f07e',
+      },
+      {
+        title: '백준 1323 복습',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: 'skyblue',
+      },
     ],
   },
   {
@@ -124,23 +233,45 @@ export const mockCalendarData = [
     month: '9',
     date: '28',
     data: [
-      { item: '코딩테스트 리뷰', backgroundColor: '#c6aca5' },
-      { item: '백준 5232 복습', backgroundColor: '#aaa5c6' },
+      {
+        title: '코딩테스트 리뷰',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: '#c6aca5',
+      },
+      {
+        title: '백준 5232 복습',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: '#aaa5c6',
+      },
     ],
   },
   {
     year: '2022',
     month: '9',
     date: '29',
-    data: [{ item: '큐 다시보기', backgroundColor: '#7ef0e3' }],
+    data: [
+      {
+        title: '큐 다시보기',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: '#7ef0e3',
+      },
+    ],
   },
   {
     year: '2022',
     month: '10',
     date: '7',
     data: [
-      { item: '데이터1', backgroundColor: '#7bc261' },
-      { item: '데이터2', backgroundColor: '#61c2ba' },
+      {
+        title: '데이터1',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: '#7bc261',
+      },
+      {
+        title: '데이터2',
+        content: 'ㅁㄴ엄ㄴ암ㄴ옴나오만왐ㄴ오마ㅓㄴ옴나오머',
+        backgroundColor: '#61c2ba',
+      },
     ],
   },
 ];
